// server/src/viewer/interface/search-line.interface.ts
export interface IResultSearchLine {
  Result: {
    IdCode: string;
    Validation: boolean;
    Name: string;
    Family: string;
    FatherName: string;
    Gender: string;
    NationalId: string;
    Serial: string;
    BirthDate: string;
    Life: boolean;
    Guardianship: boolean;
    Photo: string;
    PostalCode: string;
    Province: string;
    City: string;
    Address: string;
  };
}